Based on Playwright test feedback moved few sections from page and imported back.
